[tlo-parsing] Cleanup tlo-parsing lib interface headers.
