+++ common/tlo-parsing/flat-optimization.cpp
@@ -2,14 +2,13 @@
 // Copyright (c) 2020 LLC «V Kontakte»
 // Distributed under the GPL v3 License, see LICENSE.notice.txt
 
-#include "common/tlo-parsing/flat-optimization.h"
-
 #include <cassert>
 #include <cstring>
 #include <stdexcept>
 #include <unordered_set>
 
 #include "common/tlo-parsing/tl-objects.h"
+#include "common/tlo-parsing/tlo-parsing.h"
 
 namespace vk {
 namespace tl {
