--- common/tlo-parsing/tl-scheme-final-check.h
@@ -1,15 +1,0 @@
-// Compiler for PHP (aka KPHP)
-// Copyright (c) 2020 LLC «V Kontakte»
-// Distributed under the GPL v3 License, see LICENSE.notice.txt
-
-#pragma once
-
-namespace vk {
-namespace tl {
-
-struct tl_scheme;
-
-void tl_scheme_final_check(const tl_scheme &scheme);
-
-} // namespace tl
-} // namespace vk
