+++ compiler/tl-classes.cpp
@@ -4,9 +4,7 @@
 
 #include "compiler/tl-classes.h"
 
-#include "common/tlo-parsing/flat-optimization.h"
-#include "common/tlo-parsing/replace-anonymous-args.h"
-#include "common/tlo-parsing/tl-scheme-final-check.h"
+#include "common/tlo-parsing/tlo-parsing.h"
 #include "common/wrappers/fmt_format.h"
 
 #include "compiler/stage.h"
